feat: This commit contains the test and implementation of the wildcard host handling in `Includes`
I thought about the best way to do it, and came up with the decision to **not** duplicate code.
This required the breakdown of the `parse` function, and the creation of a function that would handle **ONLY** the extraction of host data.
Without default values and such. It has a boolean parameter to return either only the virtual or actual hosts.